Microservices/FetchingResources/requirements.txt
Flask
flask-cors
yt-dlp
moviepy
vosk
Pillow
pytesseract
fpdf
google-generativeai
dotenv
werkzeug
